[60] - Corregido el formato de integer a strings
