+++ app/Services/GetUserService.php
@@ -4,28 +4,35 @@ namespace App\Services;
 
 use App\Infrastructure\Clients\ApiClient;
 use App\Infrastructure\Clients\DBClient;
+use App\Providers\TwitchTokenProvider;
 use Illuminate\Http\JsonResponse;
 
 class GetUserService
 {
     protected ApiClient $apiClient;
     protected DBClient $dbClient;
+    private TwitchTokenProvider $tokenProvider;
 
-    public function __construct(DBClient $dbClient, ApiClient $apiClient)
+
+    public function __construct(DBClient $dbClient, ApiClient $apiClient = null)
     {
+        $this->tokenProvider = $tokenProvider ?? new TwitchTokenProvider();
+        $this->apiClient = $apiClient ?? new ApiClient($tokenProvider);
         $this->dbClient  = $dbClient;
-        $this->apiClient = $apiClient;
     }
 
+    /**
+     * @throws \Exception
+     */
     public function execute(String $userId)
     {
-        $tokenFromTwitch = $this->dbClient->getTokenFromTwitch();
+        $tokenFromTwitch = $this->tokenProvider->getTokenFromTwitch();
 
         if (!$tokenFromTwitch) {
             return response()->json(['error' => 'No se pudo obtener el token de acceso desde Twitch.'], 500);
         }
 
-        $userData = $this->apiClient->fetchUserDataFromTwitch($tokenFromTwitch, $userId);
+        $userData = $this->apiClient->fetchUserDataFromTwitch($userId);
 
         if ($userData) {
             $this->dbClient->insertUserToDB($userData);
@@ -37,8 +44,6 @@ class GetUserService
 
             return response()->json($userData, 200, [], JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES);
         }
-
         return response()->json(['error' => 'No se encontraron datos de usuario para el ID proporcionado.'], 404);
-
     }
 }
