+++ app/Services/TopsofthetopsService.php
@@ -2,10 +2,9 @@
 
 namespace App\Services;
 
+use App\Infrastructure\Clients\DBClient;
 use App\Models\TopGame;
-use App\Models\TopOfTheTop;
-use App\Models\TopVideo;
-use Carbon\Carbon;
+use App\Providers\TwitchTokenProvider;
 use Exception;
 
 /**
@@ -17,11 +16,15 @@ class TopsofthetopsService
     protected $topVideosService;
 
     protected $topGamesService;
-    public function __construct(TwitchTokenService $twitchTokenService, TopVideoService $topVideosService, TopGamesService $topGamesService)
+    protected DBClient $dbClient;
+
+    public function __construct(DBClient $dbClient = null, TwitchTokenProvider $twitchTokenService, TopVideoService $topVideosService, TopGamesService $topGamesService)
     {
         $this->twitchTokenService = $twitchTokenService;
         $this->topVideosService   = $topVideosService;
         $this->topGamesService    = $topGamesService;
+        $this->dbClient  = $dbClient;
+
     }
 
     public function updateTopOfTheTops($since)
@@ -36,54 +39,10 @@ class TopsofthetopsService
         $games = TopGame::all();
 
         foreach ($games as $game) {
-            if ($this->needsUpdate($game->game_id, $since)) {
+            if ($this->dbClient->needsUpdate($game->game_id, $since)) {
                 $this->topVideosService->updateTopVideos($game->game_id);
-                $this->updateTopForGame($game);
+                $this->dbClient->updateTopForGame($game);
             }
         }
-    }
-
-    protected function needsUpdate($gameId, $since)
-    {
-        $topOfTheTop = TopOfTheTop::find($gameId);
-
-        if (!$topOfTheTop || !$topOfTheTop->ultima_actualizacion) {
-            return true;
-        }
-
-        $lastUpdate = Carbon::parse($topOfTheTop->ultima_actualizacion);
-        $now        = Carbon::now();
-
-        return $now->diffInSeconds($lastUpdate) > $since;
-    }
-
-    protected function updateTopForGame($game)
-    {
-        $videos = TopVideo::where('game_id', $game->game_id)
-            ->orderByDesc('views')
-            ->get();
-
-        if ($videos->isEmpty()) {
-            return;
-        }
-
-        $totalViews      = $videos->sum('views');
-        $totalVideos     = $videos->count();
-        $mostViewedVideo = $videos->first();
-
-        $topOfTheTop = TopOfTheTop::updateOrCreate(
-            ['game_id' => $game->game_id],
-            [
-                'game_name'              => $game->game_name,
-                'user_name'              => $mostViewedVideo->user_name,
-                'total_videos'           => $totalVideos,
-                'total_views'            => $totalViews,
-                'most_viewed_title'      => $mostViewedVideo->title,
-                'most_viewed_views'      => $mostViewedVideo->views,
-                'most_viewed_duration'   => $mostViewedVideo->duration,
-                'most_viewed_created_at' => $mostViewedVideo->created_at,
-                'ultima_actualizacion'   => Carbon::now()
-            ]
-        );
     }
 }
