+++ routes/api.php
@@ -15,4 +15,4 @@ Route::get('/streams', StreamsController::class);
 
 Route::get('/users', UserController::class);
 
-Route::get('/topsofthetops', [TopsofthetopsController::class, 'index']);
+Route::get('/topsofthetops', TopsofthetopsController::class);
